+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+* 🆕 Supported region config.
+* 🆕 Supported language config.
+
 ## 0.1.2
 * 🆕 new parameter showSectionContents. Can be set to auto, always and onlyAfterZoom.